refactor(team): extract the OpenCode aggregate launch rollback from the aggregate run
TeamProvisioningOpenCodeAggregateRun.ts sat at 763 lines against the 800-line
default source-size limit while doing two separable jobs: driving an aggregate
OpenCode launch, and undoing one that did not survive.

Move the rollback family into TeamProvisioningOpenCodeAggregateRunRollback.ts:
the lane stop-and-rollback itself, its two progress publishers, the
untracked-lane retention helpers, the stop-confirmation assertion, and the
runtime-tracking deletion. The bodies are unchanged; the two functions the
launch flow still calls become exports.

The seam is the rollback rather than an arbitrary line count, so every function
that reads or writes runtime tracking during teardown stays together and the
launch flow is left with the two entry points it exports. That also gives the
aggregate run the headroom the following behavior changes need.
